+++ packages/shared/src/catalog/index.ts
@@ -10,6 +10,7 @@ export * from "./primitives/colorField";
 export * from "./primitives/dateField";
 export * from "./primitives/fileTrigger";
 export * from "./primitives/form";
+export * from "./primitives/gridList";
 export * from "./primitives/link";
 export * from "./primitives/listBox";
 export * from "./primitives/numberField";